% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProviders.R
\name{getProviders}
\alias{getProviders}
\title{A function to retrieve mobile broadband provider information}
\usage{
getProviders(lat, lon, zip, fips, time)
}
\arguments{
\item{lat}{The latitude of the location of interest; optional if a zipcode is provided}

\item{zip}{The zipcode of the location of interest as a character string; an optional parameter when coordinates are not known}

\item{fips}{The county-level fips code of the location of interest as a numeric string with no leading zeros; an optional parameter when coordinates are not known}

\item{time}{The segment of data of interest; check the API documentation for options}

\item{long}{The longitude of the location of interest; optional if a zipcode is provided}
}
\value{
A data frame of mobile broadband providers in a given location along with relevant technology performance metrics.
}
\description{
A function for interfacing with the national broadband map API.
}
\examples{
getProviders(lat = 39.958520, lon = -75.198857, time = "jun2014")
getProviders(zip = "19104", time = "jun2014")
getProviders(fips = 42101, time = "jun2014")
}
\keyword{API}
\keyword{broadband,}
\keyword{mobile}
