% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zips.R
\docType{data}
\name{zips}
\alias{zips}
\title{Latitude and longitude for U.S. zipcodes}
\format{A data frame with 33144 rows and 3 variables:
\describe{
\item{zip}{zipcode, as a character}
\item{lat}{latitude, as a numeric}
\item{lon}{longitude, as a numeric}
}}
\source{
\url{https://gist.github.com/erichurst/7882666}
}
\usage{
zips
}
\description{
A dataset containing the latitude and longitude coordinates for U.S. 
zipcodes.
}
\keyword{datasets}
