% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProviders.R
\name{getProviders}
\alias{getProviders}
\title{A function to retrieve mobile broadband provider information}
\usage{
getProviders(lat, lon, zip, time)
}
\arguments{
\item{lat}{The latitude of the location of interest}

\item{zip}{The zipcode of the location of interest; an optional parameter when coordinates are not known}

\item{time}{The segment of data of interest; check the API documentation for options}

\item{long}{The longitude of the location of interest}
}
\value{
A data frame of mobile broadband providers in a given location along with relevant technology performance metrics.
}
\description{
A function for interfacing with the national broadband map API.
}
\examples{
getProviders(39.958520, -75.198857, "jun2014")
getProviders("19104", "jun2014")
}
\keyword{API}
\keyword{broadband,}
\keyword{mobile}
